backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
base58==2.1.0
bip32==4.0
bitarray==2.9.3
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
ckzg==2.0.1
click==8.1.3
coincurve==20.0.0
colorama==0.4.6
cryptoconditions==0.8.1
cryptography==3.4.7
cytoolz==1.0.0
Deprecated==1.2.14
ECPy==1.2.5
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.0.0
eth_abi==5.1.0
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
h11==0.14.0
hexbytes==1.2.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mnemonic==0.21
multidict==6.0.4
openai==0.27.2
parsimonious==0.10.0
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.21.1
PyNaCl==1.4.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pyunormalize==16.0.0
pywin32==308
regex==2024.9.11
requests==2.28.2
rlp==4.0.1
six==1.16.0
sniffio==1.3.1
toolz==1.0.0
tqdm==4.65.0
types-Deprecated==1.2.9.20240311
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.12.2
urllib3==1.26.15
web3==7.3.1
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.16.0
xrpl-py==3.0.0
yarl==1.8.2
zipp==3.15.0
